readme and cleanup
