+++ example/README.md
@@ -2,9 +2,19 @@
 
 This example demonstrates how to use the `groth16-solana` crate to verify a proof generated by the `SP1` library on Solana.
 
-## Usage: Script
+## Overview: Script
 
-To run the example, you can use the following ELF options:
+This script will load a [`SP1ProofWithPublicValues`](https://docs.rs/sp1-sdk/2.0.0/sp1_sdk/proof/struct.SP1ProofWithPublicValues.html)
+from the pre-generated proofs in the [`proofs`](./proofs) directory and verify them. Optionally, you can 
+pass in the `--prove` flag to generate a fresh proof from the provided [`elfs`](./elfs) and save it to the `proofs` directory.
+
+The loaded `SP1ProofWithPublicValues` will be serialized into a `SP1ProofFixture`, and written to the
+[`proof-fixtures`](./proof-fixtures) directory. The `SP1ProofFixture` can easily be borsh serialized and 
+deserialized, and is used to verify the proof. 
+
+### Running the script
+
+To run the example script, you can use the following ELF options:
 
 | ELF Option | Description                   |
 | ---------- | ----------------------------- |
@@ -21,8 +31,7 @@ cargo build
 RUST_LOG=info cargo run --release --  --elf fibonacci 
 ```
 
-The solana program 
-Pass in the `--prove` flag to generate a fresh proof and save it to a file:
+Pass in the `--prove` flag to generate a fresh proof from the provided ELF and save it to a file:
 
 ```shell
 cd script
@@ -33,6 +42,15 @@ RUST_LOG=info cargo run --release -- --elf fibonacci --prove
 Note: The pre-generated proof for `fibonacci` assumes the input `n` is 20. The pre-generated proof 
 for `is-prime` assumes the input `n` is 11.
 
-## Usage: Solana Program
+## Overview: Solana Program
 
-The solana program 
+The code in [`solana/program`](./solana/program) is a simple Solana program that verifies a `SP1ProofFixture` using the `groth16-solana` crate.
+It costs roughly 234000 compute units. 
+
+### Running the solana program
+
+```shell
+cd solana
+npm install
+yarn build-and-test
+```
