+++ example/proof-fixtures/fibonacci_fixture.bin
