--- example/solana/ts/index.ts
@@ -1,4 +1,0 @@
-import { PublicKey } from '@solana/web3.js';
-export * from './instructions';
-
-export const PROGRAM_ID = new PublicKey('Fg6PaFpoGXkYsidMpWTK6W2BeZ7FEfcYkg476zPFsLnS');
